Import OpenSim geometry files
Geometry files, also sometimes referred to as bone files, are xml files
with a .vtk file extension.

This addon is under development and not yet functional.  It is certainly
not robust.  It is not tolerant of errors.  It will even allow the user
to select a file that is not an OpenSim geometry file.