Permuted y and z axes. Plus, changed to be addon.
In OpenSim y is up.  In Blender z is up.  To conform to Blender's
convention vertices in OpenSim geometry files are altered as follows
Blender x = OpenSim x
Blender z = OpenSim y
Blender y = - OpenSim z

This commit also commented out and uncommented a few lines so that the
script could be incorporated as an addon.  I'm not doing this quite
right so these will change.  For now the addon is functional.