Now using proper addon mechanics.
In the previous version, I was mixing code for handling panel buttons (I
think) with code for adding a menu item to the File > Import menu.

I found a great Blender Import Guide that had a template for setting up
an import addon.  That template exists in the ExamplesTemplate directory
and is called ImportTemplate.py.  The source is noted in the template.

So, now, the OpenSimImportGeometry script is functioning as a proper
addon.  It does not crash when no geometry is selected (so got rid of a
major bug), and it does not need lines commented in/out when running as
a script vs. when installed as an addon.